requirements.txt
pytest
playwright
pytest-playwright
pytest-html
allure-pytest